redirect output with StringIO to print with click.echo()
