+++ cli/environment.py
@@ -1,9 +1,11 @@
 from deploybot.environment import Environment
 from .config import Config
-
+from . import response
+from io import StringIO
 import click
 import tableprint
 
+output = StringIO()
 pass_config = click.make_pass_decorator(Config, ensure=True)
 
 
@@ -20,10 +22,10 @@ def environment_list(config, repository=None):
     client = Environment(config.client)
     result = client.list(repository)
     header = ('ID', 'Repository ID', 'Environment Name', 'Branch', 'Automatic', 'Current')
-    # data = response('environment', 'list', result)
+    data = response('environment', 'list', result)
 
-    # tableprint.table(data, headers=header, width=int(config.width), style=config.style)
-    click.echo(result)
+    tableprint.table(data, headers=header, width=int(config.width), style=config.style, out=output)
+    click.echo(output.getvalue())
 
 
 @environment.command('get')
@@ -34,5 +36,7 @@ def environment_get(config, id):
     client = Environment(config.client)
     result = client.get(id)
     header = ('ID', 'Repository ID', 'Environment Name', 'Branch', 'Automatic', 'Current')
+    data = response('environment', 'get', result)
 
-    click.echo(result)
+    tableprint.table(data, headers=header, width=int(config.width), style=config.style, out=output)
+    click.echo(output.getvalue())
